+++ .eslintrc.js
@@ -5,6 +5,7 @@ module.exports = {
     'jsx-a11y/click-events-have-key-events': 'warn',
     'jsx-a11y/no-static-element-interactions': 'warn',
 
+    'smarthr/a11y-anchor-has-href-attribute': 'error',
     'smarthr/a11y-prohibit-input-placeholder': 'error',
   },
 }