+++ src/components/DropZone/DropZone.tsx
@@ -5,12 +5,14 @@ import React, {
   forwardRef,
   useCallback,
   useImperativeHandle,
+  useMemo,
   useRef,
   useState,
 } from 'react'
 import styled, { css } from 'styled-components'
 
 import { Theme, useTheme } from '../../hooks/useTheme'
+import { DecoratorsType } from '../../types/props'
 import { Button } from '../Button'
 import { FaFolderOpenIcon } from '../Icon'
 
@@ -35,7 +37,11 @@ type DropZoneProps = {
   multiple?: boolean
   children?: React.ReactNode
   name?: string
+  /** コンポーネント内の文言を変更するための関数を設定 */
+  decorators?: DecoratorsType<'selectButtonLabel'>
 }
+
+const SELECT_BUTTON_LABEL = 'ファイルを選択'
 
 const overrideEventDefault = (e: DragEvent<HTMLElement>) => {
   e.preventDefault()
@@ -43,7 +49,7 @@ const overrideEventDefault = (e: DragEvent<HTMLElement>) => {
 }
 
 export const DropZone = forwardRef<HTMLInputElement, DropZoneProps & ElementProps>(
-  ({ children, onSelectFiles, accept, multiple = true, name }, ref) => {
+  ({ children, onSelectFiles, accept, multiple = true, name, decorators }, ref) => {
     const theme = useTheme()
     const classNames = useClassNames()
     const fileRef = useRef<HTMLInputElement>(null)
@@ -52,6 +58,11 @@ export const DropZone = forwardRef<HTMLInputElement, DropZoneProps & ElementProp
     useImperativeHandle<HTMLInputElement | null, HTMLInputElement | null>(
       ref,
       () => fileRef.current,
+    )
+
+    const selectButtonLabel = useMemo(
+      () => decorators?.selectButtonLabel?.(SELECT_BUTTON_LABEL) || SELECT_BUTTON_LABEL,
+      [decorators],
     )
 
     const onDrop = useCallback(
@@ -97,7 +108,7 @@ export const DropZone = forwardRef<HTMLInputElement, DropZoneProps & ElementProp
       >
         {children}
         <Button prefix={<FaFolderOpenIcon />} onClick={onClickButton}>
-          ファイルを選択
+          {selectButtonLabel}
         </Button>
         <input
           ref={fileRef}