+++ src/components/Dropdown/DropdownMenuButton/DropdownMenuButton.stories.tsx
@@ -27,12 +27,12 @@ const Template: React.FC<Omit<ComponentProps<typeof DropdownMenuButton>, 'childr
 
 export const Default: Story = () => (
   <Cluster align="center">
-    <Template />
-    <Template disabled />
-    <Template onlyIconTrigger />
+    <Template label="その他の操作" />
+    <Template disabled label="その他の操作" />
+    <Template onlyIconTrigger label="その他の操作" />
     <Template triggerSize="s" label="操作" />
     <Template triggerSize="s" label={<span>操作</span>} disabled />
-    <Template triggerSize="s" onlyIconTrigger />
+    <Template triggerSize="s" onlyIconTrigger label="操作" />
   </Cluster>
 )
 Default.storyName = 'DropdownMenuButton'